Render a ViewComponent
Some light refactoring to support rendering a ViewComponent specified by a reflex.

At the moment we rely on the Reflex setting the component instance variable but this could be tidied up into a module or reflex superclass.

We add a new method render_component which uses the controller’s view_context to render the component to a string.

I extracted the standard morph generation (for each selector scope, extract the partial html from the full html) into a collect_morphs method. Since the component is rendering a partial update already, we assign the component output to each selector.