Extract rendering functionality to a Renderer
Allow Reflexes to control rendering abilities via the morphs method which calculates the changes the required to be broadcast back.

In this implementation, rendering is just a module that's included in the Reflex. By default we include the ControllerRenderer but I've also sketched out a ComponentRenderer. This flexibility means the method of rendering is not confined to what the core StimulusReflex code provides and allows the reflex implementation itself to control the rendering.

The renderer is primarily a `morphs` method which, given a list of selectors, returns a nested array of [selector, html] instances. The Channel therefore isn't responsible for generating these morphs—it's the renderer's job to produce the html and how it maps to a selector.

In the case of a ControllerRenderer, it renders a whole page and pulls out the selectors specified. In the case of a ComponentRenderer, it render the same component html for each selector.

These renderers could just as easily be class instances attached to a reflex (allowing the reflex to still override them or provide a different implementation). Not wedded to the module approach—it was just simpler for right now.

Importantly: none of the javascript logic has changed and the rendering method doesn't leak into the html. It's all contained in the reflex

For example, a component reflex just looks like

```ruby
class PartialReflex < StimulusReflex::Reflex
  include StimulusReflex::ComponentRenderer

  def random
  end

  def component
    FastRandomComponent.new
  end
end
```

```html
  <a href="#" data-reflex="click->PartialReflex#random" data-reflex-root="#fast">Partial refresh</a>
```

(where #fast is the id of the component div to update)